Add arachne GUI module and main script to run
This commit adds a module that creates a GUI to display two images
side-by-side along with captions. The GUI isn't particular resilient as
it doesn't respond well to being resized (a.k.a. the wrong size!), but
this probably isn't worth addressing as it's intended to run full screen
on a screen of known size.

This commit also introduces some checks in the form of a GitHub Action
to ensure Python codebase quality.